+++ cx-portal/src/components/pages/AdminBoardDetail/BoardContentDetails.tsx
@@ -31,6 +31,7 @@ import BoardProvider from './components/BoardProvider'
 import { AppDetails, DocumentTypeText } from 'features/apps/apiSlice'
 import './AdminBoardDetail.scss'
 import CommonService from 'services/CommonService'
+import BoardPrivacy from './components/BoardPrivacy'
 
 export default function BoardContentDetails({ item }: { item: AppDetails }) {
   const { t } = useTranslation()
@@ -56,6 +57,7 @@ export default function BoardContentDetails({ item }: { item: AppDetails }) {
         {images && <BoardImageGallery images={images} />}
         <BoardConnectedData item={item} />
         <BoardSecurityInfo />
+        <BoardPrivacy item={item} />
         <BoardDocuments
           type={DocumentTypeText.CONFORMITY_DOCUMENT}
           appId={item.id}