+++ CHANGELOG.md
@@ -6,6 +6,10 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 
 * AppReleaseProcess
    * enabled document download in contract and consent
+* AppDetails
+   * Privacy Policy
+* AdminBoardDetail
+   * Privacy Policy
 
 ### Change